requirements.txt
Django==1.9.8
mysqlclient==1.3.7
djangorestframework==3.4.0
markdown==2.6.6
django-filter==0.13.0
djangorestframework-jwt==1.8.0
django-cors-headers==1.1.0